fix@metrics_pipeline.com/environment/requirements.txt
flask==2.3.0
werkzeug==3.0.0
pyyaml==6.0.1
requests==2.31.0
